requirements.txt
python-telegram-bot==20.8
psycopg2-binary==2.9.9
aiohttp==3.9.1
stripe==7.9.0